refactor components directories
