+++ components/LoadingComponent/Spinner.vue
